+++ src/serverMiddleware/routes.js
@@ -3,7 +3,7 @@
 // using the ./proxy.js serverMiddleware
 
 // const id = "[a-z0-9]+";
-// const mongoId = "[a-z0-9]{24}";
+const mongoId = "[a-z0-9]{24}";
 const activationCode = "[a-z0-9]+";
 // const courseId = "[a-f0-9]{24}";
 const routes = [
@@ -34,7 +34,7 @@ const routes = [
 	//`^/news`,
 	`^/news/new`,
 	//`^/news/${mongoId}/?$`,
-	//`^/news/${mongoId}/edit`,
+	`^/news/${mongoId}/edit`,
 	`^/nuxtversion/?$`,
 	// `^/courses/?$`,
 ];