+++ src/components/organisms/HomeworksList.unit.js
@@ -5,7 +5,6 @@ import {
 	openHomeworks,
 } from "@@/stories/mockData/Homeworks";
 import Vuetify from "vuetify";
-import { fromNowToFuture } from "@plugins/datetime";
 
 describe("@components/organisms/HomeworksList", () => {
 	const mockStore = {
@@ -33,6 +32,17 @@ describe("@components/organisms/HomeworksList", () => {
 
 	it(...isValidComponent(HomeworksList));
 
+	it("accepts only student and teacher as type prop", () => {
+		const validTypes = ["student", "teacher"];
+		const { validator } = HomeworksList.props.type;
+
+		validTypes.forEach((type) => {
+			expect(validator(type)).toBe(true);
+		});
+
+		expect(validator("wrong type")).toBe(false);
+	});
+
 	it("Should render complete homework items list", () => {
 		const wrapper = mount(HomeworksList, {
 			...createComponentMocks({
@@ -43,6 +53,7 @@ describe("@components/organisms/HomeworksList", () => {
 			vuetify,
 			propsData: {
 				homeworks,
+				type: "student",
 			},
 		});
 
@@ -76,43 +87,12 @@ describe("@components/organisms/HomeworksList", () => {
 				store: mockStoreEmpty,
 			}),
 			vuetify,
-		});
-
-		expect(wrapper.props("homeworks")).toStrictEqual([]);
-		expect(wrapper.findAllComponents({ name: "VListItem" })).toHaveLength(0);
-	});
-
-	it("Should link list item links to homework/<id> page", () => {
-		const wrapper = mount(HomeworksList, {
-			...createComponentMocks({
-				i18n: true,
-				vuetify: true,
-				store: mockStore,
-			}),
-			vuetify,
 			propsData: {
-				homeworks,
+				type: "student",
 			},
 		});
-
-		const firstLink = wrapper.find("a");
-
-		expect(firstLink.exists()).toBe(true);
-		expect(firstLink.attributes().href).toBe(`/homework/${homeworks[0]._id}`);
-	});
-
-	it("Should display due date labels according to due date", () => {
-		const wrapper = mount(HomeworksList, {
-			...createComponentMocks({
-				i18n: true,
-				vuetify: true,
-				store: mockStore,
-			}),
-			vuetify,
-			propsData: {
-				homeworks,
-			},
-		});
+		// expect(wrapper.props("homeworks")).toStrictEqual([]);
+		// expect(wrapper.findAllComponents({ name: "VListItem" })).toHaveLength(0);
 
 		const dueDateLabels = wrapper.findAll("[data-test-id='dueDateLabel']");
 		expect(dueDateLabels).toHaveLength(homeworks.length);
@@ -229,6 +209,7 @@ describe("@components/organisms/HomeworksList", () => {
 			vuetify,
 			propsData: {
 				homeworks: [],
+				type: "student",
 			},
 		});
 
@@ -238,5 +219,19 @@ describe("@components/organisms/HomeworksList", () => {
 		).toBe(true);
 		expect(wrapper.props("homeworks")).toStrictEqual([]);
 		expect(wrapper.findAllComponents({ name: "VListItem" })).toHaveLength(0);
+	});
+
+	it("should accept valid type props", () => {
+		const { validator } = HomeworksList.props.type;
+		const validTypes = ["student", "teacher"];
+		const invalidTypes = ["invalid", "type"];
+
+		validTypes.forEach((type) => {
+			expect(validator(type)).toBe(true);
+		});
+
+		invalidTypes.forEach((type) => {
+			expect(validator(type)).toBe(false);
+		});
 	});
 });